Created a database test widget
This widget allows you to enter a sql query into the textfield and will execute it and display the result when the "execute" button is pressed. We can use this as a starting point for interacting with the database but I think that we will need to make some changes to make sure that sql injection attacks are not viable.